refactoring classes and implemented base abstract class
