+++ aiohttp_rate_limiter/limiters/sliding_log.py
@@ -1,35 +1,13 @@
-from typing import Callable
-from asyncio import get_event_loop, sleep
+from aiohttp.web_middlewares import middleware
 
-from aiohttp.web import middleware
-
-from ..cfg import Config
+from aiohttp_rate_limiter.limiters.base import BaseLimiter
 
 
-class SlidingLog:
-    def __init__(self, config: Config, error_handler: Callable):
-        self._config: Config = config
-
-        self._clients = {}
-
-        self._error_handler: Callable = error_handler
-
-        self._check_typing()
-        self._set_interval()
-
+class SlidingLog(BaseLimiter):
     def _check_typing(self):
         cfg = self._config
         assert type(cfg.max_requests) == int
         assert type(cfg.interval) == int
-
-    def _set_interval(self):
-        async def clear_count(self):
-            while True:
-                self._clients = {}
-                await sleep(self._config.interval)
-
-        loop = get_event_loop()
-        loop.create_task(clear_count(self))
 
     @middleware
     async def handle(self, request, handler):
